Rename update
Fixed the sub-folder routine by implementing a new class. Added a browser to scan different folders. Merged Ogle and Fill methods. Added a feature to check for the correct extension. Changed the location of .rename to the Documents folder and added an auto-close option, in case you don't want the application to close after committing to changes.